+++ control/__init__.py
@@ -60,7 +60,7 @@ from .phaseplot import phase_plot, box_grid
 from .pzmap import pzmap
 from .rlocus import root_locus
 from .statefbk import place, lqr, ctrb, obsv, gram, acker
-from .statesp import StateSpace
+from .statesp import *
 from .timeresp import forced_response, initial_response, step_response, \
     impulse_response
 from .xferfcn import TransferFunction
@@ -86,7 +86,7 @@ except ImportError:
 #! of defaults from the main package.  At that point, the matlab module will
 #! allow provide compatibility with MATLAB but no package functionality.
 #!
-from .matlab import ss, tf, ss2tf, tf2ss, drss
+from .matlab import tf, ss2tf
 from .matlab import pole, zero, evalfr, freqresp, dcgain
 from .matlab import nichols, rlocus, margin
         # bode and nyquist come directly from freqplot.py
